pyproject.toml
[tool.poetry]
name = "code-leap"
version = "0.1.0"
description = ""
authors = ["Every"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
